analysis/requirements.txt
# =============================================================================
# Honeypot Threat Intelligence Platform — Python Dependencies
# Install with: pip install -r requirements.txt
# =============================================================================

# ── Core data handling
pandas>=2.1.0
numpy>=1.26.0

# ── Elasticsearch client
elasticsearch>=8.10.0

# ── Threat intelligence enrichment APIs
shodan>=1.30.0
abuseipdb-wrapper>=0.1.1
vt-py>=0.18.0

# ── GeoIP & geographic analysis
geoip2>=4.7.0
geopandas>=0.14.0
shapely>=2.0.0

# ── STIX 2.1 / TAXII
stix2>=3.0.1
taxii2-client>=2.3.0
cabby>=0.1.23

# ── Machine learning (clustering)
scikit-learn>=1.3.0

# ── Visualisation
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0

# ── Networking / HTTP
requests>=2.31.0
urllib3>=2.0.7
dnspython>=2.4.2
ipwhois>=1.2.0

# ── Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
rich>=13.6.0
loguru>=0.7.2
click>=8.1.7
pyyaml>=6.0.1

# ── MISP integration
pymisp>=2.4.177

# ── TAXII server (optional)
fastapi>=0.104.0
uvicorn>=0.24.0
